+++ CHANGELOG.md
@@ -16,8 +16,11 @@ These are the section headers that we use:
 
 ## [Unreleased]()
 
+## [1.21.0](https://github.com/argilla-io/argilla/compare/v1.20.0...v1.21.0)
+
 ### Added
 
+- Added new draft queue for annotation view ([#4334](https://github.com/argilla-io/argilla/pull/4334))
 - Added annotation metrics module for the `FeedbackDataset` (`argilla.client.feedback.metrics`). ([#4175](https://github.com/argilla-io/argilla/pull/4175)).
 - Added strategy to handle and translate errors from the server for `401` HTTP status code` ([#4362](https://github.com/argilla-io/argilla/pull/4362))
 - Added integration for `textdescriptives` using `TextDescriptivesExtractor` to configure `metadata_properties` in `FeedbackDataset` and `FeedbackRecord`. ([#4400](https://github.com/argilla-io/argilla/pull/4400)). Contributed by @m-newhauser
@@ -28,6 +31,7 @@ These are the section headers that we use:
 
 ### Changed
 
+- More productive and simpler shortcuts system ([#4215](https://github.com/argilla-io/argilla/pull/4215))
 - Move `ArgillaSingleton`, `init` and `active_client` to a new module `singleton`. ([#4347](https://github.com/argilla-io/argilla/pull/4347))
 - Updated `argilla.load` functions to also work with `FeedbackDataset`s. ([#4347](https://github.com/argilla-io/argilla/pull/4347))
 - [breaking] Updated `argilla.delete` functions to also work with `FeedbackDataset`s. It now raises an error if the dataset does not exist. ([#4347](https://github.com/argilla-io/argilla/pull/4347))
@@ -37,6 +41,10 @@ These are the section headers that we use:
 
 - Fixed error in `TextClassificationSettings.from_dict` method in which the `label_schema` created was a list of `dict` instead of a list of `str`. ([#4347](https://github.com/argilla-io/argilla/pull/4347))
 - Fixed total records on pagination component ([#4424](https://github.com/argilla-io/argilla/pull/4424))
+
+### Removed
+
+- Removed `draft` auto save for annotation view ([#4334](https://github.com/argilla-io/argilla/pull/4334))
 
 ## [1.20.0](https://github.com/argilla-io/argilla/compare/v1.19.0...v1.20.0)